Splitted buildSearch and meta into separate file.
